Fixxed compatibility issue
Both broker Mod_Gearman and Statusengin uses „client“ as variable name
if for the gearman_client. If both brokers are loaded by nagios they
get a part of the same application. The same name of the variable
brings a strange behavior to the system.

Install.sh with colors now
Gearman server can be set in Config/Statusengine.php for the php app now